Update pump schema
Remove attributes from mandatory list that do not have do be specified
or cannot be specified simultaneously.

Remove `head` attribute, as if a `head` is specified in the EPANET
network model this is a head curve. Rename the `pumpCurve` attribute
as `headCurve`.

Remove `headCurve` attribute from example, as either `headCurve` or
`power` is required, not both.